fix gui package
