+++ instances/bukkit/build.gradle.kts
@@ -39,7 +39,7 @@ dependencies {
     implementation(projects.modules.integrationPapi)
     implementation(projects.modules.core.api)
     implementation(projects.modules.core.bukkit)
-    implementation(projects.modules.guiCoreBukkit)
+    implementation(projects.modules.core.guiBukkit)
     implementation(projects.modules.commandBukkit)
     implementation(projects.modules.eventBukkit)
     implementation(projects.modules.ratingChange.api)
