+++ modules/command-bukkit/src/main/kotlin/ru/astrainteractive/astrarating/command/di/CommandsModule.kt
@@ -4,20 +4,20 @@ import CommandManager
 import ru.astrainteractive.astralibs.lifecycle.Lifecycle
 import ru.astrainteractive.astrarating.core.di.BukkitModule
 import ru.astrainteractive.astrarating.core.di.CoreModule
+import ru.astrainteractive.astrarating.core.gui.di.GuiBukkitModule
 import ru.astrainteractive.astrarating.feature.rating.change.di.RatingChangeModule
-import ru.astrainteractive.astrarating.gui.di.GuiModule
 
 class CommandsModule(
     ratingChangeModule: RatingChangeModule,
     bukkitModule: BukkitModule,
     coreModule: CoreModule,
-    guiModule: GuiModule
+    guiBukkitModule: GuiBukkitModule
 ) {
     private val commandManager: CommandManager by lazy {
         val dependencies = CommandsDependencies.Default(
             bukkitModule = bukkitModule,
             coreModule = coreModule,
-            guiModule = guiModule,
+            guiBukkitModule = guiBukkitModule,
             ratingChangeModule = ratingChangeModule
         )
         CommandManager(dependencies)
