+++ modules/command-bukkit/src/main/kotlin/ru/astrainteractive/astrarating/command/rating/RatingCommandExecutor.kt
@@ -9,11 +9,11 @@ import org.bukkit.entity.Player
 import ru.astrainteractive.astralibs.command.api.executor.CommandExecutor
 import ru.astrainteractive.astralibs.kyori.KyoriComponentSerializer
 import ru.astrainteractive.astralibs.permission.BukkitPermissibleExt.toPermissible
+import ru.astrainteractive.astrarating.core.gui.router.GuiRouter
 import ru.astrainteractive.astrarating.core.settings.AstraRatingTranslation
 import ru.astrainteractive.astrarating.data.exposed.dto.RatingType
 import ru.astrainteractive.astrarating.data.exposed.model.PlayerModel
 import ru.astrainteractive.astrarating.feature.rating.change.domain.usecase.AddRatingUseCase
-import ru.astrainteractive.astrarating.gui.router.GuiRouter
 import ru.astrainteractive.klibs.mikro.core.dispatchers.KotlinDispatchers
 
 internal class RatingCommandExecutor(
