+++ modules/core/gui-bukkit/build.gradle.kts
