+++ modules/core/gui-bukkit/src/main/kotlin/ru/astrainteractive/astrarating/core/gui/util/TranslationDesc.kt
@@ -1,0 +1,8 @@
+package ru.astrainteractive.astrarating.core.gui.util
+
+import ru.astrainteractive.astralibs.string.StringDesc
+import ru.astrainteractive.astrarating.core.settings.AstraRatingTranslation
+
+internal class TranslationDesc(private val provider: (AstraRatingTranslation) -> StringDesc.Raw) {
+    fun toString(astraRatingTranslation: AstraRatingTranslation) = provider.invoke(astraRatingTranslation)
+}
