+++ modules/event-bukkit/build.gradle.kts
@@ -22,5 +22,5 @@ dependencies {
     implementation(projects.modules.core.bukkit)
     implementation(projects.modules.data.dao)
     implementation(projects.modules.data.exposed)
-    implementation(projects.modules.guiCoreBukkit)
+    implementation(projects.modules.core.guiBukkit)
 }
