+++ modules/rating-players/api/src/main/kotlin/ru/astrainteractive/astrarating/feature/rating/players/di/RatingPlayersModule.kt
@@ -3,9 +3,9 @@ package ru.astrainteractive.astrarating.feature.rating.players.di
 import kotlinx.coroutines.CoroutineScope
 import ru.astrainteractive.astralibs.lifecycle.Lifecycle
 import ru.astrainteractive.astrarating.data.dao.di.RatingDaoModule
+import ru.astrainteractive.astrarating.feature.rating.players.data.RatingPlayersCachedRepositoryImpl
 import ru.astrainteractive.astrarating.feature.rating.players.presentation.DefaultRatingPlayersComponent
 import ru.astrainteractive.astrarating.feature.rating.players.presentation.RatingPlayersComponent
-import ru.astrainteractive.astrarating.feature.rating.players.data.RatingPlayersCachedRepositoryImpl
 import ru.astrainteractive.klibs.mikro.core.dispatchers.KotlinDispatchers
 
 class RatingPlayersModule(
