+++ settings.gradle.kts
@@ -51,6 +51,6 @@ include("modules:rating-player:api")
 include("modules:rating-players:api")
 include("modules:core:api")
 include("modules:core:bukkit")
-include("modules:gui-core-bukkit")
+include("modules:core:gui-bukkit")
 include("modules:command-bukkit")
 include("modules:event-bukkit")
